refactor : change paths and naming for folders
