+++ src/layouts/index.ts
@@ -1,2 +1,3 @@
 export { Layout } from './layout/Layout'
 export { LayoutWithSidebar } from './layoutWithSidebar/LayoutWithSidebar'
+export { getLayoutWithSidebar } from './layoutWithSidebar/LayoutWithSidebar'
