+++ src/pages/index.tsx
@@ -1,11 +1,16 @@
 import { useRouter } from 'next/router'
 import { useEffect } from 'react'
+import Cookies from "js-cookie";
 
 export default function Page() {
   const router = useRouter()
 
   useEffect(() => {
-    router.push('/usersList')
+    if(!Cookies.get('authToken')){
+      router.push('/auth/login')
+    }else{
+      router.push('/usersList')
+    }
   }, [router])
 
   return null
